Update 11_lesson.rst
The total_cost value requires a String as input in order to work correctly, however, in the tutorial the default value is given as an int. 

In order for the exercises to run quotation marks should be added as in the tutorials_11.zip file that you are able to download.